Clojure: Add more tools.
Former-commit-id: 1171da033d51e7382f65c74c5d569028afae4dc8

+++ emacs/setup.sh
@@ -9,7 +9,9 @@ fi
 ln -sf "$SCRIPT_DIR/spacemacs" ~/.spacemacs
 
 # Systemd service
+[[ -f ~/.config/systemd/user/emacs.service ]] && exit 0
 mkdir -p ~/.config/systemd/user
 ln -sf "$SCRIPT_DIR/emacs.service" ~/.config/systemd/user/emacs.service
-# systemctl --user enable emacs
-# systemctl --user start  emacs
+systemctl --user daemon-reload
+systemctl --user enable emacs
+systemctl --user start  emacs
